chore: move enum folder
